% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adbcpostgresql-package.R
\name{adbcpostgresql}
\alias{adbcpostgresql}
\alias{adbc_database_init.adbcpostgresql_driver_postgresql}
\alias{adbc_connection_init.adbcpostgresql_database}
\alias{adbc_statement_init.adbcpostgresql_connection}
\title{ADBC PostgreSQL Driver}
\usage{
adbcpostgresql()

\method{adbc_database_init}{adbcpostgresql_driver_postgresql}(driver, ..., uri)

\method{adbc_connection_init}{adbcpostgresql_database}(database, ..., adbc.connection.autocommit = NULL)

\method{adbc_statement_init}{adbcpostgresql_connection}(
  connection,
  ...,
  adbc.ingest.target_table = NULL,
  adbc.ingest.target_db_schema = NULL,
  adbc.ingest.mode = NULL
)
}
\arguments{
\item{driver}{The driver to use. This can be one of the following:
\itemize{
\item A non-missing \code{character(1)} containing a driver or manifest name, a
relative or absolute path to a driver or manifest, or a URI. For a URI,
the driver manager uses the URI scheme as the driver name and passes the
URI to that driver. A \verb{profile://} URI loads a connection profile.
\item \code{NULL}, which leaves driver selection to the driver manager. In this
case, \code{...} must normally contain \code{uri} or \code{profile}; the driver is
inferred from the URI or loaded from the connection profile.
\item An object that inherits from the \code{adbc_driver} class, such as one created
by \code{\link[adbcdrivermanager:adbc_driver]{adbc_driver()}}. This includes drivers provided by R packages, such as
\code{adbcsqlite::adbcsqlite()}.
}}

\item{...}{Driver-specific options. These are generally named values that
are converted to strings.}

\item{uri}{A URI to a database path (e.g.,
\verb{postgresql://localhost:1234/postgres?user=user&password=password})}

\item{database}{An \link[adbcdrivermanager:adbc_database_init]{adbc_database}.}

\item{adbc.connection.autocommit}{Use FALSE to disable the default
autocommit behaviour.}

\item{connection}{An \link[adbcdrivermanager:adbc_connection_init]{adbc_connection}}

\item{adbc.ingest.target_table}{The name of the target table for a bulk insert.}

\item{adbc.ingest.target_db_schema}{The schema of the table for a bulk insert.}

\item{adbc.ingest.mode}{Whether to create (the default) or append.}
}
\value{
An \code{\link[adbcdrivermanager:adbc_driver]{adbcdrivermanager::adbc_driver()}}
}
\description{
ADBC PostgreSQL Driver
}
\examples{
adbcpostgresql()

}
